Adds support for Xen VTPM provisioning
These changes allow Nova to create VTPM stub domains for compute instances. To
support this we request a UUID for the VTPM from Xen and create a small 2MB
backing image and create a VTPM stub via libvirt using these resources.

Since this is just a proof of concept we have made a number of assertions
tailored to our particular deployment.